For_PostgreSQL/pyproject.toml
[tool.poetry]
name = "for-postgresql"
version = "0.1.0"
description = ""
authors = ["Serhii-SA <serg.a.tolstov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^24.4.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
